Add unit test. Switch from CURL to Guzzle to allow this.
I have added an initial unit test, switching to the guzzle library which provides support for this. Once
fully switched to Guzzle CURL is not longer required as a dependency as Guzzle will use it if available,
but can procede without it.

Adding testability in this way also permits libraries that use this to incorporate mock results in
their own tests. You can configure https://travis-ci.org/ to run the unit tests on any pull request.