Check xml validity coming back from silverpop.
In some cases (related to start_time & end_time being too close I think) silver pop returns invalid xml.

In this case we should throw an exception as current behaviour is more of a silent fail.

I'm not a big fan of the method I used but I didn't want to go down the xml library replacement
path this round